requirements.txt
aiohttp==1.0.5
asn1crypto==0.22.0
async-timeout==1.2.1
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
discord.py==0.16.10
idna==2.6
keyring==10.4.0
multidict==3.1.3
pycparser==2.18
requests==2.18.4
SecretStorage==2.3.1
six==1.10.0
urllib3==1.22
websockets==3.4
wikipedia==1.4.0
